# api/openapi.yaml
openapi: 3.1.0
info:
  title: AirShield Privacy Control Plane API
  version: 1.0.0
  description: Tenant-isolated English PII/PHI/PCI protection, signed evidence, identity binding, deletion, and
    dual-control reidentification. Production callers authenticate with workload OIDC JWTs. Source code alone is
    not certification or compliance evidence.
paths:
...
  /v1/protect:
    post:
      summary: Protect Route
      operationId: protect_route_v1_protect_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProtectRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProtectResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Missing or invalid workload identity
        '403':
          description: Tenant or scope authorization denied
      description: Protect text and issue a signed, hash-chained, metadata-only egress receipt.
      x-required-scope: airshield.protect
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProtectRequest:
      properties:
        session_id:
          type: string
          pattern: ^ses_[a-f0-9]{24}$
          title: Session Id
        text:
          type: string
          minLength: 1
          title: Text
        policy:
          type: string
          pattern: ^[a-z0-9][a-z0-9-]{2,127}$
          title: Policy
        destination:
          type: string
          pattern: ^[a-z][a-z0-9-]{2,63}$
          title: Destination
        speaker_token:
          anyOf:
          - type: string
            pattern: ^\[SPEAKER_[A-F0-9]{6}\]$
          - type: 'null'
          title: Speaker Token
        idempotency_key:
          type: string
          maxLength: 128
          minLength: 16
          pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]+$
          title: Idempotency Key
      type: object
      required:
      - session_id
      - text
      - policy
      - destination
      - idempotency_key
      title: ProtectRequest
    ProtectResponse:
      properties:
        protected_text:
          type: string
          title: Protected Text
        entities:
          items:
            $ref: '#/components/schemas/ProtectedEntity'
          type: array
          title: Entities
        decision:
          type: string
          enum:
          - allow
          - block
          - review
          title: Decision
        receipt:
          $ref: '#/components/schemas/ReceiptResponse'
      type: object
      required:
      - protected_text
      - entities
      - decision
      - receipt
      title: ProtectResponse
    ProtectedEntity:
      properties:
        type:
          type: string
          title: Type
        token:
          type: string
          title: Token
        start:
          type: integer
          title: Start
        end:
          type: integer
          title: End
        confidence:
          type: number
          title: Confidence
      type: object
      required:
      - type
      - token
      - start
      - end
      - confidence
      title: ProtectedEntity
    ReceiptResponse:
      properties:
        receipt_id:
          type: string
          title: Receipt Id
        sequence:
          type: integer
          title: Sequence
        content_sha256:
          type: string
          title: Content Sha256
        policy:
          type: string
          title: Policy
        policy_version:
          type: string
          title: Policy Version
        entity_counts:
          additionalProperties:
            type: integer
          type: object
          title: Entity Counts
        destination:
          type: string
          title: Destination
        decision:
          type: string
          title: Decision
        reason_codes:
          items:
            type: string
          type: array
          title: Reason Codes
        previous_hash:
          type: string
          title: Previous Hash
        event_hash:
          type: string
          title: Event Hash
        signature:
          type: string
          title: Signature
        signing_key_id:
          type: string
          title: Signing Key Id
        signing_algorithm:
          type: string
          title: Signing Algorithm
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - receipt_id
      - sequence
      - content_sha256
      - policy
      - policy_version
      - entity_counts
      - destination
      - decision
      - previous_hash
      - event_hash
      - signature
      - signing_key_id
      - signing_algorithm
      - created_at
      title: ReceiptResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
      bearerFormat: OIDC JWT
      description: Trusted workload JWT; a verified signed claim scopes every tenant operation.
servers:
- url: https://airshield-control.internal
  description: Private customer deployment